Compilation error removed: "Unnecessary cast from int to int"

+++ spikes/src/spikes/rene/toscoball/Mesa.java
@@ -192,7 +192,7 @@ public class Mesa extends JPanel {
 	}
 	
 	public void space(int num) {
-		dist=(int)Math.min(250,Math.max(18,dist+num));
+		dist = Math.min(250,Math.max(18,dist+num));
 	}
 	
 	public void shoot() {
